+++ MoerderCode/app/src/main/java/com/example/jendrik/moerder/GUI/OnGamingClasses/Suspect.java
@@ -135,7 +135,7 @@ public class Suspect extends Fragment {
      */
     public void onClickSuspect(){
         String player= (String)spinnerPlayer.getSelectedItem();
-        Suspection sus = new Suspection(getActivity(), player, room, weapon, game, game.getActivePlayer().getpNumber());
+        Suspection sus = new Suspection(getActivity(), player, room, weapon, game, (int)game.getActivePlayer().getpNumber());
         //TODO irgendwie ein endTurn() einbauen
     }
